TST: test_solidmotor.py reorganized.
